VPNaaS: A collection tracker for container scheduler
A CollectionTracker is an ObjectTracker abstract class that includes
functionality for tracking collection of objects, where each object
is represented by its own ObjectTracker. This class merges the
notifications from all child trackers into a single notification
that is emitted on the output observable as:

   Map(child_id -> last_child_event)

The class provides three methods that allow callers to specify the
child objects that need to be tracked: `watch`, `add` and `remove`.
The `watch` method updates the list of tracked objects with the
given one. The method computes the difference between the current
tracked objects and the new identfiers, and only updates the
children as needed. The methods `add` and `remove` allow discrete
modifications of the current set of tracked objects.

Classes extending the collection tracker must implement the
`newMember` method which returns a new child ObjectTracker for a
given child identifier.

Change-Id: Id0bcfb04fb83d5d398e9f3a861c606aa0815094a
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>