VPNaaS: A host group tracker for container scheduler
This is an ObjectTracker that monitors the updates for a given
host group and emits notifications with the status of the
container service running on all hosts in the group.

Internally, the host group tracker uses a CollectionTracker to
receive the notifications from the group members and exposes these
notifications as a single stream of updates.

Change-Id: Ie3ddb7069a20c191d490cef83afdd8668537b1f4
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>